+++ xdef/src/main/java/org/xdef/impl/code/DefXmlWriter.java
@@ -28,12 +28,10 @@ import java.math.BigInteger;
 public class DefXmlWriter extends KXmlOutStream
 	implements XDXmlOutStream, XDValue {
 
-	@SuppressWarnings("deprecation")
 	/** Creates new instance of DefXmlOutStream with java.io.Writer
-	 * @deprecated please use OutputStream instead of Writer.
 	 * @param writer where to write XML.
 	 * @param encoding encoding of XML stream.
-	 * @param writeDocumentHeader if <tt>true</tt> then the XML header is
+	 * @param writeDocumentHeader if true then the XML header is
 	 * written, otherwise no XML header is written.
 	 */
 	public DefXmlWriter(final Writer writer,